requirements.txt
pyyaml~=6.0.2
nicegui~=2.11.0
pydantic~=2.10.6